refactor: 커뮤니티 메인 fsd 구조로 리팩토링 (#61)
- 퀴즈 배너 아직 수정 안함
- 다른 페이지 관련 파일들은 정리하지 않은 상태
