+++ src/app/page.tsx
@@ -1,20 +1,11 @@
-import {
-  CommunityBanner,
-  CommunityFilters,
-  PostCard,
-  FloatingButton,
-} from '@/features/community/ui'
-import { MOCK_POSTS } from '@/features/community/mockData'
-import UrlPagination from '@/shared/ui/UrlPagination'
+import CommunityBanner from '@/widgets/community-quiz/ui/CommunityBanner'
+import CommunityBoard from '@/widgets/community-board/ui/CommunityBoard'
+import FloatingButton from '@/shared/ui/FloatingButton'
+import { CommunityBoardSearchParams } from '@/widgets/community-board/model/types'
 
 // nuqs 쓰면 거기서 다시 처리
 interface PageProps {
-  searchParams: Promise<{
-    page: string
-    category: 'all' | 'free' | 'recruit' | 'study'
-    sort: 'popular' | 'latest'
-    query: string
-  }>
+  searchParams: Promise<CommunityBoardSearchParams>
 }
 
 export default async function Page({ searchParams }: PageProps) {
@@ -27,38 +18,12 @@ export default async function Page({ searchParams }: PageProps) {
       <CommunityBanner />
 
       {/* 게시판 */}
-      <section className="mt-16 space-y-8">
-        <h1 className="text-brand-black text-4xl font-extrabold">게시판</h1>
-
-        {/* 게시판 헤더 */}
-        <CommunityFilters
-          activeCategory={category ?? 'all'}
-          sortBy={sort ?? 'popular'}
-          searchParams={{ category, sort, query, page }}
-        />
-
-        {/* 게시글 목록 */}
-        <ul
-          id="community-post-list"
-          className="flex flex-col gap-4 border-b-2 pb-8"
-        >
-          {MOCK_POSTS.length > 0 ? (
-            MOCK_POSTS.map((post) => <PostCard key={post.id} post={post} />)
-          ) : (
-            <div className="text-brand-gray-300 py-20 text-center">
-              해당 게시글이 없습니다.
-            </div>
-          )}
-        </ul>
-
-        {/* 페이지네이션 */}
-        <UrlPagination
-          page={Number(page) || 1}
-          totalPages={50} // TODO: API 수정 요청함. 결과에 따라 처리.
-          searchParams={{ page, category, sort, query }}
-          className="mb-16 py-4"
-        />
-      </section>
+      <CommunityBoard
+        page={page}
+        category={category}
+        sort={sort}
+        query={query}
+      />
 
       {/* 플로팅 버튼 */}
       <FloatingButton />
