+++ src/entities/post/model/thumbnail1.jpg
