+++ src/entities/post/model/thumbnail2.jpg
