+++ src/entities/quiz/bannerMockData.ts
