+++ src/entities/quiz/model/schema.ts
