+++ src/features/community/assets/index.ts
@@ -1,27 +1,6 @@
-import profile1 from './profile1.jpg'
-import profile2 from './profile2.png'
-import profile3 from './profile3.jpg'
-import profile4 from './profile4.jpg'
-import thumbnail1 from './thumbnail1.jpg'
-import thumbnail2 from './thumbnail2.jpg'
-import thumbnail3 from './thumbnail3.jpg'
-import thumbnail4 from './thumbnail4.jpg'
 import QuizImage from './quiz-image.png'
 import QuizImageWide from './quiz-image-wide.png'
 import QuoteImage from './quote-image.png'
 import QuoteImageWide from './quote-image-wide.png'
 
-export {
-  profile1,
-  profile2,
-  profile3,
-  profile4,
-  thumbnail1,
-  thumbnail2,
-  thumbnail3,
-  thumbnail4,
-  QuizImage,
-  QuizImageWide,
-  QuoteImage,
-  QuoteImageWide,
-}
+export { QuizImage, QuizImageWide, QuoteImage, QuoteImageWide }
