+++ src/features/community/ui/index.ts
@@ -1,6 +1,3 @@
-import CommunityBanner from '@/features/community/ui/CommunityBanner'
-import CommunityFilters from '@/features/community/ui/CommunityFilters'
-import PostCard from '@/features/community/ui/PostCard'
-import FloatingButton from '@/features/community/ui/FloatingButton'
+import CommunityBanner from '@/widgets/community-quiz/ui/CommunityBanner'
 
-export { CommunityBanner, CommunityFilters, PostCard, FloatingButton }
+export { CommunityBanner }
