+++ src/shared/ui/FloatingButton.tsx
